+++ nuxt.config.ts
@@ -129,7 +129,7 @@ export default defineNuxtConfig({
   },
   colorMode: {
     classSuffix: "",
-    preference: "dark",
+    preference: "system",
   },
   vite: {
     build: {